UPC SKU search

+++ src/main/java/com/bedbath/jdalookups/model/SkuOrUpcSearchReq.java
@@ -15,8 +15,8 @@ public class SkuOrUpcSearchReq {
 	private String region;
 	private String lifeStyle;
 	private String itemStatus;
-	private Long skuNumber;
-	private Long upcNumber;
+	private long skuNumber;
+	private long upcNumber;
 	private String itemDescription;
 	private int vendor;
 	private String partNumber;
@@ -28,12 +28,12 @@ public class SkuOrUpcSearchReq {
 	private int listNumber;
 	private String merchandiseGroup;
 	private String priceGroup;
-	private int retailFrom;
-	private int retailTo;
-	private int mapFrom;
-	private int mapTo;
-	private int msrpFrom;
-	private int msrpTo;
+	private float retailFrom;
+	private float retailTo;
+	private float mapFrom;
+	private float mapTo;
+	private float msrpFrom;
+	private float msrpTo;
 	private String buyer;
 	private String exclusivity;
 	private String chefCentral;
@@ -43,247 +43,488 @@ public class SkuOrUpcSearchReq {
 	private int limit;
 	private String server;
 	
+
+	
 	public String getType() {
 		return type;
 	}
+
+
+
 	public void setType(String type) {
 		this.type = type;
 	}
+
+
+
 	public int getDepartment() {
 		return department;
 	}
+
+
+
 	public void setDepartment(int department) {
 		this.department = department;
 	}
+
+
+
 	public int getSubDepartment() {
 		return subDepartment;
 	}
+
+
+
 	public void setSubDepartment(int subDepartment) {
 		this.subDepartment = subDepartment;
 	}
+
+
+
 	public int getClassa() {
 		return classa;
 	}
+
+
+
 	public void setClassa(int classa) {
 		this.classa = classa;
 	}
+
+
+
 	public int getColorCode() {
 		return colorCode;
 	}
+
+
+
 	public void setColorCode(int colorCode) {
 		this.colorCode = colorCode;
 	}
+
+
+
 	public String getSizeCode() {
 		return sizeCode;
 	}
+
+
+
 	public void setSizeCode(String sizeCode) {
 		this.sizeCode = sizeCode;
 	}
+
+
+
 	public String getWebProduct() {
 		return webProduct;
 	}
+
+
+
 	public void setWebProduct(String webProduct) {
 		this.webProduct = webProduct;
 	}
+
+
+
 	public String getWebCollection() {
 		return webCollection;
 	}
+
+
+
 	public void setWebCollection(String webCollection) {
 		this.webCollection = webCollection;
 	}
+
+
+
 	public String getSeason() {
 		return season;
 	}
+
+
+
 	public void setSeason(String season) {
 		this.season = season;
 	}
+
+
+
 	public String getTheme() {
 		return theme;
 	}
+
+
+
 	public void setTheme(String theme) {
 		this.theme = theme;
 	}
+
+
+
 	public String getRegion() {
 		return region;
 	}
+
+
+
 	public void setRegion(String region) {
 		this.region = region;
 	}
+
+
+
 	public String getLifeStyle() {
 		return lifeStyle;
 	}
+
+
+
 	public void setLifeStyle(String lifeStyle) {
 		this.lifeStyle = lifeStyle;
 	}
+
+
+
 	public String getItemStatus() {
 		return itemStatus;
 	}
+
+
+
 	public void setItemStatus(String itemStatus) {
 		this.itemStatus = itemStatus;
 	}
-	public Long getSkuNumber() {
+
+
+
+	public long getSkuNumber() {
 		return skuNumber;
 	}
-	public void setSkuNumber(Long skuNumber) {
+
+
+
+	public void setSkuNumber(long skuNumber) {
 		this.skuNumber = skuNumber;
 	}
-	public Long getUpcNumber() {
+
+
+
+	public long getUpcNumber() {
 		return upcNumber;
 	}
-	public void setUpcNumber(Long upcNumber) {
+
+
+
+	public void setUpcNumber(long upcNumber) {
 		this.upcNumber = upcNumber;
 	}
+
+
+
 	public String getItemDescription() {
 		return itemDescription;
 	}
+
+
+
 	public void setItemDescription(String itemDescription) {
 		this.itemDescription = itemDescription;
 	}
+
+
+
 	public int getVendor() {
 		return vendor;
 	}
+
+
+
 	public void setVendor(int vendor) {
 		this.vendor = vendor;
 	}
+
+
+
 	public String getPartNumber() {
 		return partNumber;
 	}
+
+
+
 	public void setPartNumber(String partNumber) {
 		this.partNumber = partNumber;
 	}
+
+
+
 	public String getVdc() {
 		return vdc;
 	}
+
+
+
 	public void setVdc(String vdc) {
 		this.vdc = vdc;
 	}
+
+
+
 	public String getVendorCollection() {
 		return vendorCollection;
 	}
+
+
+
 	public void setVendorCollection(String vendorCollection) {
 		this.vendorCollection = vendorCollection;
 	}
+
+
+
 	public String getVendorStyle() {
 		return vendorStyle;
 	}
+
+
+
 	public void setVendorStyle(String vendorStyle) {
 		this.vendorStyle = vendorStyle;
 	}
+
+
+
 	public String getEcomBrand() {
 		return ecomBrand;
 	}
+
+
+
 	public void setEcomBrand(String ecomBrand) {
 		this.ecomBrand = ecomBrand;
 	}
+
+
+
 	public String getProductType() {
 		return productType;
 	}
+
+
+
 	public void setProductType(String productType) {
 		this.productType = productType;
 	}
+
+
+
 	public int getListNumber() {
 		return listNumber;
 	}
+
+
+
 	public void setListNumber(int listNumber) {
 		this.listNumber = listNumber;
 	}
+
+
+
 	public String getMerchandiseGroup() {
 		return merchandiseGroup;
 	}
+
+
+
 	public void setMerchandiseGroup(String merchandiseGroup) {
 		this.merchandiseGroup = merchandiseGroup;
 	}
+
+
+
 	public String getPriceGroup() {
 		return priceGroup;
 	}
+
+
+
 	public void setPriceGroup(String priceGroup) {
 		this.priceGroup = priceGroup;
 	}
-	public int getRetailFrom() {
+
+
+
+	public float getRetailFrom() {
 		return retailFrom;
 	}
-	public void setRetailFrom(int retailFrom) {
+
+
+
+	public void setRetailFrom(float retailFrom) {
 		this.retailFrom = retailFrom;
 	}
-	public int getRetailTo() {
+
+
+
+	public float getRetailTo() {
 		return retailTo;
 	}
-	public void setRetailTo(int retailTo) {
+
+
+
+	public void setRetailTo(float retailTo) {
 		this.retailTo = retailTo;
 	}
-	public int getMapFrom() {
+
+
+
+	public float getMapFrom() {
 		return mapFrom;
 	}
-	public void setMapFrom(int mapFrom) {
+
+
+
+	public void setMapFrom(float mapFrom) {
 		this.mapFrom = mapFrom;
 	}
-	public int getMapTo() {
+
+
+
+	public float getMapTo() {
 		return mapTo;
 	}
-	public void setMapTo(int mapTo) {
+
+
+
+	public void setMapTo(float mapTo) {
 		this.mapTo = mapTo;
 	}
-	public int getMsrpFrom() {
+
+
+
+	public float getMsrpFrom() {
 		return msrpFrom;
 	}
-	public void setMsrpFrom(int msrpFrom) {
+
+
+
+	public void setMsrpFrom(float msrpFrom) {
 		this.msrpFrom = msrpFrom;
 	}
-	public int getMsrpTo() {
+
+
+
+	public float getMsrpTo() {
 		return msrpTo;
 	}
-	public void setMsrpTo(int msrpTo) {
+
+
+
+	public void setMsrpTo(float msrpTo) {
 		this.msrpTo = msrpTo;
 	}
+
+
+
 	public String getBuyer() {
 		return buyer;
 	}
+
+
+
 	public void setBuyer(String buyer) {
 		this.buyer = buyer;
 	}
+
+
+
 	public String getExclusivity() {
 		return exclusivity;
 	}
+
+
+
 	public void setExclusivity(String exclusivity) {
 		this.exclusivity = exclusivity;
 	}
+
+
+
 	public String getChefCentral() {
 		return chefCentral;
 	}
+
+
+
 	public void setChefCentral(String chefCentral) {
 		this.chefCentral = chefCentral;
 	}
+
+
+
 	public String getProfile() {
 		return profile;
 	}
+
+
+
 	public void setProfile(String profile) {
 		this.profile = profile;
 	}
+
+
+
 	public String getEph() {
 		return eph;
 	}
+
+
+
 	public void setEph(String eph) {
 		this.eph = eph;
 	}
+
+
+
 	public int getStart() {
 		return start;
 	}
+
+
+
 	public void setStart(int start) {
 		this.start = start;
 	}
+
+
+
 	public int getLimit() {
 		return limit;
 	}
+
+
+
 	public void setLimit(int limit) {
 		this.limit = limit;
 	}
+
+
+
 	public String getServer() {
 		return server;
 	}
+
+
+
 	public void setServer(String server) {
 		this.server = server;
 	}
-	
+
+
+
 	@Override
 	public String toString() {
 		return "SkuOrUpcSearchReq [type=" + type + ", department=" + department + ", subDepartment=" + subDepartment
